more documentations and abstractions
